requirements.txt
gradio>=4.36
pandas
numpy
phonenumbers
huggingface_hub>=0.23
langchain>=0.2
langchain-community>=0.2
langchain-huggingface>=0.0.3
pydantic>=2
python-dotenv
openai>=1.43.0